Fixed Value Conversion Warnings (file size)

+++ Observer/Maps/Map.h
@@ -33,7 +33,7 @@
 @property (nonatomic, strong, readonly) NSString *description;
 @property (nonatomic, strong, readonly) NSString *author;
 @property (nonatomic, strong, readonly) NSDate *date;
-@property (nonatomic, readonly) NSUInteger byteCount;
+@property (nonatomic, readonly) unsigned long long byteCount;
 @property (nonatomic, readonly) AGSEnvelope *extents;
 @property (nonatomic, strong, readonly) NSURL *thumbnailUrl;
